Updated the sdk to patch the dynamic link tables in it's entry point, this means that symbols like fread/fwrite/fseek link correctly and we should be able to link with normal libs without re-compiling.

+++ trunk/dingoo_sdk/src/libc/stddef.h
@@ -1,14 +1,12 @@
 #ifndef __stddef_h__
 #define __stddef_h__
 
-#include <stdint.h>
-
 #ifndef NULL
 #define NULL ((void *)0)
 #endif
 
-typedef intptr_t ptrdiff_t;
-typedef uintptr_t size_t;
-typedef uint16_t wchar_t;
+typedef long int ptrdiff_t;
+typedef unsigned long int size_t;
+typedef unsigned short wchar_t;
 
 #endif
